DockerProject/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pillow>=9.0.0
soundfile>=0.12.0
numpy>=1.24.0
requests>=2.28.0
icrawler>=0.6.0
torch>=1.13.0
kokoro 